Try to consume the entire 407 response to prevent BadStatusLine because of leftovers
Since some NTLM proxies send HTLM pages after the blank line that
determines the 407 request is over, we need to continue reading in order
to flush the socket file so the next response starts where it should.
However, other NTLM proxies do not send this HTLM page so we need to use
`select` to determine when we can read more from the response and when to exit right away.